+++ src/main/java/lotto/service/DBLoader.java
@@ -1,6 +1,9 @@
 package lotto.service;
 
-import lotto.dao.*;
+import lotto.dao.LottoDAO;
+import lotto.dao.LottoResultDAO;
+import lotto.dao.RoundDAO;
+import lotto.dao.WinningLottoDAO;
 import lotto.domain.LottoResult;
 import lotto.domain.Lottos;
 import lotto.domain.WinningLotto;
@@ -8,27 +11,7 @@ import lotto.domain.WinningLotto;
 import java.sql.Connection;
 import java.sql.SQLException;
 
-public class DBAccessor {
-
-    public static Lottos getDBLottos(Connection connection, String lottoRound) throws SQLException {
-        LottoDAO lottoDAO = new LottoDAO(connection);
-        return lottoDAO.findByLottoRound(lottoRound);
-    }
-
-    public static WinningLotto getDBWinningLotto(Connection connection, String lottoRound) throws SQLException {
-        WinningLottoDAO winningLottoDAO = new WinningLottoDAO(connection);
-        return winningLottoDAO.findByLottoRound(lottoRound);
-    }
-
-    public static LottoResultDTO getDBLottoResultDTO(Connection connection, String lottoRound) throws SQLException {
-        LottoResultDAO lottoResultDAO = new LottoResultDAO(connection);
-        return lottoResultDAO.findByLottoRound(lottoRound);
-    }
-
-    public static Integer getNextRound(Connection connection) throws SQLException {
-        RoundDAO roundDAO = new RoundDAO(connection);
-        return roundDAO.getNextRound();
-    }
+public class DBLoader {
 
     public static void loadDBRoundTable(Connection connection, Integer round) throws SQLException {
         RoundDAO roundDAO = new RoundDAO(connection);
